requirements.txt
django==4.2.7
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-cors-headers